requirements.txt
pyyaml
mysql-connector-python
yfinance
ta